--- js/data/pictures.js
@@ -1,14 +1,0 @@
-const pictures = {
-  paintings: [
-    `https://k42.kn3.net/CF42609C8.jpg`,
-    `https://k42.kn3.net/D2F0370D6.jpg`,
-    `https://k32.kn3.net/5C7060EC5.jpg`
-  ],
-  photos: [
-    `http://i.imgur.com/1KegWPz.jpg`,
-    `https://i.imgur.com/DiHM5Zb.jpg`,
-    `http://i.imgur.com/DKR1HtB.jpg`
-  ]
-};
-
-export default pictures;
